Convert to proper structure
