+++ functions/src/api/helloWorld.ts
@@ -1,6 +1,6 @@
 import { Request } from 'firebase-functions/lib/providers/https';
 import { Response } from 'express-serve-static-core';
-import { firebase, functions } from './shared';
+import { firebase, functions } from '../shared';
 
 export const helloWorld = functions.https.onRequest((request: Request, response: Response) => {
   firebase.logger.info('Hello logs!', { structuredData: true });
