+++ functions/src/cron/index.ts
@@ -1,0 +1,1 @@
+export { ingestUnityVersions } from './ingestUnityVersions';
