+++ functions/src/index.ts
@@ -1,5 +1,3 @@
-import * as buildQueue from './buildQueue';
-import * as ingest from './ingest';
-import { helloWorld } from './helloWorld';
-
-export { helloWorld, ingest, buildQueue };
+export * as model from './model';
+export * as cron from './cron';
+export * from './api';
