--- functions/src/ingest.ts
@@ -1,31 +1,0 @@
-import { Request } from 'firebase-functions/lib/providers/https';
-import { Response } from 'express-serve-static-core';
-import { EventContext } from 'firebase-functions';
-import { firebase, functions } from './shared';
-import { getUnityVersionInfo } from './logic/getUnityVersionInfo';
-
-export const intoBrowser = functions.https.onRequest(
-  async (request: Request, response: Response) => {
-    try {
-      const versionInfo = await getUnityVersionInfo();
-      response.send(versionInfo);
-    } catch (err) {
-      firebase.logger.error(err);
-      response.send('Oops.');
-    }
-  },
-);
-
-export const intoDb = functions.pubsub
-  .schedule('every 5 minutes')
-  .onRun(async (context: EventContext) => {
-    try {
-      const versionInfo = await getUnityVersionInfo();
-
-      // Todo - ingest in db
-
-      firebase.logger.info(versionInfo);
-    } catch (err) {
-      firebase.logger.error(err);
-    }
-  });
