+++ functions/src/model/buildQueue.ts
@@ -1,0 +1,15 @@
+import { EventContext } from 'firebase-functions';
+import { QueryDocumentSnapshot } from 'firebase-functions/lib/providers/firestore';
+import { firebase, functions } from '../shared';
+
+export const onItemCreated = functions.firestore
+  .document('buildQueue/{itemId}')
+  .onCreate((snap: QueryDocumentSnapshot, context: EventContext) => {
+    // Todo - create repository_dispatch event
+    // https://developer.github.com/v3/repos/#create-a-repository-dispatch-event
+
+    // Todo - Lock record on successful dispatch
+
+    const newValue = snap.data();
+    firebase.logger.info(newValue);
+  });
